git-svn-id: https://svn.eu.apache.org/repos/asf/logging/log4j/trunk@565523 13f79535-47bb-0310-9956-ffa450edef68

+++ tests/src/java/org/apache/log4j/FileAppenderTest.java
@@ -17,6 +17,8 @@
 
 package org.apache.log4j;
 
+import junit.framework.TestCase;
+
 import java.io.File;
 
 import java.lang.reflect.Method;
@@ -24,52 +26,33 @@ import java.lang.reflect.Method;
 
 /**
  *
- * Test if WriterAppender honors the Appender contract.
+ * FileAppender tests.
  *
- * @author <a href="http://www.qos.ch/log4j/">Ceki G&uuml;lc&uuml;</a>
  * @author Curt Arnold
  */
-public class FileAppenderTest extends AbstractAppenderTest {
-  protected AppenderSkeleton getAppender() {
-    return new FileAppender();
-  }
-
-  protected AppenderSkeleton getConfiguredAppender() {
-    FileAppender wa = new FileAppender();
-    wa.setFile("output/temp");
-    wa.setLayout(new DummyLayout());
-
-    return wa;
-  }
-
-  public void testPartiallyConfiguredAppender() {
-    FileAppender wa1 = new FileAppender();
-    wa1.setFile("output/temp");
-    assertFalse(wa1.isActive());
-
-    FileAppender wa2 = new FileAppender();
-    wa2.setLayout(new DummyLayout());
-    assertFalse(wa2.isActive());
-  }
-
+public class FileAppenderTest extends TestCase {
   /**
    * Tests that any necessary directories are attempted to
    * be created if they don't exist.  See bug 9150.
    *
    */
   public void testDirectoryCreation() {
-    File newFile = new File("output/newdir/temp.log");
-    newFile.delete();
+    //
+    //   known to fail on JDK 1.1
+    if (!System.getProperty("java.version").startsWith("1.1.")) {
+      File newFile = new File("output/newdir/temp.log");
+      newFile.delete();
 
-    File newDir = new File("output/newdir");
-    newDir.delete();
+      File newDir = new File("output/newdir");
+      newDir.delete();
 
-    FileAppender wa = new FileAppender();
-    wa.setFile("output/newdir/temp.log");
-    wa.setLayout(new DummyLayout());
-    wa.activateOptions();
+      org.apache.log4j.FileAppender wa = new org.apache.log4j.FileAppender();
+      wa.setFile("output/newdir/temp.log");
+      wa.setLayout(new PatternLayout("%m%n"));
+      wa.activateOptions();
 
-    assertTrue(new File("output/newdir/temp.log").exists());
+      assertTrue(new File("output/newdir/temp.log").exists());
+    }
   }
 
   /**
@@ -95,20 +78,10 @@ public class FileAppenderTest extends AbstractAppenderTest {
 
   /**
    * Tests isAsSevereAsThreshold.
-   * @deprecated
    */
   public void testIsAsSevereAsThreshold() {
     FileAppender appender = new FileAppender();
     Priority debug = Level.DEBUG;
     assertTrue(appender.isAsSevereAsThreshold(debug));
-  }
-
-    /**
-     * Test for bug 38993.
-     * @throws java.io.IOException if IOException
-     */
-  public void testSetFileBuffered() throws java.io.IOException {
-      FileAppender appender = new FileAppender();
-      appender.setFile("output/setFileBuffered.log", false, true, 100);
   }
 }
